Finished the project
The idea behind this was mainly to gain experience and learn about making games, specifically about making platformers. I will do different genres in the future but it will never leave the scope of a simple learning project with no further goal than to learn.

For the changes, i added a goal (which is to get the golden triangle by collecting all orbs and squares) a primitive save and load function, for that one I wanted to do much more but I don't know how to and don't want to rely too much on simply copy pasting videos on youtube so I did the basic only.